Добавил возможность записи в топик

+++ data-managment-service/src/main/java/ru/spbstu/rakitin/management/service/impl/TopicServiceImpl.java
@@ -3,17 +3,21 @@ package ru.spbstu.rakitin.management.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.spbstu.rakitin.commonentites.model.Topic;
 import ru.spbstu.rakitin.commonstarter.admin.AdminManager;
+import ru.spbstu.rakitin.commonstarter.discovery.AdminUserService;
 import ru.spbstu.rakitin.dto.PermissionTypeEnum;
-import ru.spbstu.rakitin.management.exception.KafkaTopicCreationException;
-import ru.spbstu.rakitin.management.exception.TopicAlreadyCreatedException;
-import ru.spbstu.rakitin.management.exception.TopicNotFoundException;
-import ru.spbstu.rakitin.management.exception.TopicQuotaLimitException;
+import ru.spbstu.rakitin.dto.TaskType;
+import ru.spbstu.rakitin.dto.kafka.KafkaMessageDto;
+import ru.spbstu.rakitin.management.dto.TaskDto;
+import ru.spbstu.rakitin.management.exception.*;
 import ru.spbstu.rakitin.management.repository.TopicRepository;
 import ru.spbstu.rakitin.management.service.KafkaService;
+import ru.spbstu.rakitin.management.service.TaskService;
 import ru.spbstu.rakitin.management.service.TopicService;
 
+import java.util.Arrays;
 import java.util.List;
 
 @Service
@@ -23,7 +27,7 @@ public class TopicServiceImpl implements TopicService {
     private final TopicRepository topicRepository;
     private final AdminManager adminManager;
     private final KafkaService kafkaService;
-
+    private final TaskService taskService;
 
     @Override
     public Topic findTopicById(long id, Authentication authentication) throws TopicNotFoundException {
@@ -46,11 +50,22 @@ public class TopicServiceImpl implements TopicService {
     }
 
     @Override
-    public void deleteTopic(long topicId, Authentication authentication) throws TopicNotFoundException {
+    @Transactional
+    public void deleteTopic(long topicId, Authentication authentication) throws TopicNotFoundException, TopicDeleteForbidden {
         Topic topic = findTopicById(topicId, authentication);
         adminManager.checkAccessThrowable(authentication, topic.getProject().getId(), PermissionTypeEnum.TOPIC_CREATE);
-        kafkaService.deleteTopic(topic);
+        List<TaskDto> tasks = taskService.getTasksListForProjectsAndTaskTypes(List.of(topic.getProject().getId()), Arrays.stream(TaskType.values()).toList());
+        if(tasks.stream().anyMatch(taskDto -> taskDto.getJobDto().getTopicId() == topicId)) {
+            throw new TopicDeleteForbidden(String.format("Unable to delete topic with id %s because there are running tasks for this topic", topicId));
+        }
         topicRepository.delete(topic);
+        kafkaService.deleteTopic(topic);
+    }
+
+    @Override
+    public void sendMessageToTopic(long topicId, KafkaMessageDto message, Authentication authentication) throws TopicNotFoundException {
+        Topic topic = findTopicById(topicId, authentication);
+        kafkaService.sendMessageToTopic(topic.getNameInKafka(), message.getMessage(), message.getKey());
     }
 
 }
